Create Sign. py
This script is a digital signature tool that can be used for creating and verifying digital signatures for messages. Digital signatures are used to ensure the authenticity and integrity of messages.

The script has the following functionality:

Generating a new private/public key pair: The script uses the RSA (Rivest-Shamir-Adleman) algorithm to generate a new private and public key pair. The private key is used to sign messages and the public key is used to verify signatures.

Signing a message: Given a private key and a message, the script generates a digital signature of the message. The signature is generated by first creating a hash of the message using the SHA-256 algorithm and then signing the hash using the private key.

Verifying a signature: Given a public key, a message, and a signature, the script verifies the signature. The script first creates a hash of the message using the SHA-256 algorithm, and then verifies the signature against the hash and the public key. If the signature is valid, the script outputs "The signature is valid." If the signature is not valid, the script outputs "The signature is not valid."

The script uses the argparse library to parse command-line arguments, the hashlib library to generate hashes of messages, and the rsa library to perform RSA operations.

The script can be run by providing the appropriate command-line arguments. For example, to generate a new private/public key pair, you would run the script with the '-g' or '--generate' option. To sign a message, you would run the script with the '-s' or '--sign' option, along with the private key and the message. To verify a signature, you would run the script with the '-v' or '--verify' option, along with the public key, the message, and the signature.
Created with ChatGPT lol I give credit where it's due 
Seriously even it debugged almost all the errors